(#20) create :: studentSignUpUseCase
