+++ src/main/kotlin/kr/hs/entrydsm/exit/domain/student/presentation/StudentController.kt
@@ -1,18 +1,20 @@
 package kr.hs.entrydsm.exit.domain.student.presentation
 
-import kr.hs.entrydsm.exit.domain.student.presentation.dto.request.StudentUpdateRequest
-import kr.hs.entrydsm.exit.domain.student.usecase.StudentUpdateProfileUseCase
+import kr.hs.entrydsm.exit.domain.student.presentation.dto.request.StudentSignUpRequest
+import kr.hs.entrydsm.exit.domain.student.usecase.StudentSignUpUseCase
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
+import javax.validation.Valid
 
 @RequestMapping("/student")
 @RestController
 class StudentController(
-    private val updateProfileUseCase: StudentUpdateProfileUseCase
+    private val updateProfileUseCase: StudentSignUpUseCase
 ) {
+
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PatchMapping("/profile")
-    fun updateProfile(@RequestBody request: StudentUpdateRequest) {
+    fun updateProfile(@RequestBody @Valid request: StudentSignUpRequest) {
         updateProfileUseCase.execute(request)
     }
 }
