--- src/main/kotlin/kr/hs/entrydsm/exit/domain/student/presentation/dto/request/StudentUpdateRequest.kt
@@ -1,10 +1,0 @@
-package kr.hs.entrydsm.exit.domain.student.presentation.dto.request
-
-data class StudentUpdateRequest (
-    val name: String,
-    val profileImagePath: String,
-    val grade: String,
-    val classNum: String,
-    val number: String,
-    val major: String
-)
